+++ models/accrual.go
@@ -21,5 +21,5 @@ type OrderToAccrual struct {
 type AccrualResponse struct {
 	Order   string             `json:"order"`
 	Status  AccrualOrderStatus `json:"status"`
-	Accrual float64            `json:"accrual"`
+	Accrual float32            `json:"accrual"`
 }
