+++ models/withdraw.go
@@ -2,12 +2,12 @@ package models
 
 type WithdrawRequest struct {
 	OrderNumber string  `json:"order"`
-	Sum         float64 `json:"sum"`
+	Sum         float32 `json:"sum"`
 }
 
 type WithdrawalsResponse struct {
 	UUID        string  `json:"uuid,omitempty"`
 	OrderNumber string  `json:"order"`
-	Sum         float64 `json:"sum"`
+	Sum         float32 `json:"sum"`
 	ProcessedAt string  `json:"processed_at"`
 }
